pyproject.toml
[tool.poetry]
name = "reflexion-agent"
version = "0.1.0"
description = ""
authors = ["Gokul Gowri Shankar <gokul.gowrishankar@wavicledata.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
langchain = "^0.1.20"
langchain-openai = "^0.1.6"
langgraph = "^0.0.48"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
